Fix postData (#4)
Requests originating from specs can be malformed: the values in Rails's environment are not necessarily strings, even though the content type is purported to be "application/x-www-form-urlencoded". If we find a non-string value in any parameter, work around this by converting the post data to "application/json".

Also removed explicit `json` dependency, based on user feedback.

Bumped to v0.2.2.